move MockConsole, start to get_object decorator
