+++ paths_cli/tests/wizard/test_wizard.py
@@ -1,7 +1,7 @@
 import pytest
 import mock
 from paths_cli.tests.wizard.mock_wizard import (
-    make_mock_wizard, make_mock_retry_wizard
+    MockConsole
 )
 import pathlib
 
@@ -10,30 +10,6 @@ from openpathsampling.experimental.storage.collective_variables import \
 
 from paths_cli.wizard.wizard import *
 
-class MockConsole:
-    def __init__(self, inputs=None):
-        if isinstance(inputs, str):
-            inputs = [inputs]
-        elif inputs is None:
-            inputs = []
-        self.inputs = inputs
-        self._input_iter = iter(inputs)
-        self.log = []
-        self.width = 80
-        self.input_call_count = 0
-
-    def print(self, content=""):
-        self.log.append(content)
-
-    def input(self, content):
-        self.input_call_count += 1
-        user_input = next(self._input_iter)
-        self.log.append(content + " " + user_input)
-        return user_input
-
-    @property
-    def log_text(self):
-        return "\n".join(self.log)
 
 
 class TestWizard:
